requirements.txt
mcp==1.10.1
python-dotenv==1.0.1